Change uint64_t to uintptr_t for addresses
Until now, PBD relied on uint64_t types to store address for 64-bit
values. Since PBD will support x86 too, we should not use uint64_t
to store addresses anymore, since this could lead to subtle issues
in the future.

Regarding portability, I know that (u/i)ntptr_t is, unfortunately,
marked as optional in the C99 standard and the reasoning for that
makes some sense: some archs have pointers that is bigger than the
maximum integral type available in C, such as 128-bit pointers on
the IBM AS/400. However, I do not expect the PBD to be ported to
exotic architectures like that, and I think (u/i)ntptr_t is easily
found on many OSs, libraries, and most used archs, so let's use it.

+++ src/analysis.c
@@ -619,7 +619,7 @@ void static_analysis_finish(void)
  * does, but in a 'smarter' way.
  */
 struct hashtable *static_analysis(const char *file, const char *func,
-	struct array *lines_l, uint64_t firstbreak)
+	struct array *lines_l, uintptr_t firstbreak)
 {
 	char *file_cur;
 	struct dw_line *line;
